Updated the UI of the dashboard
