+++ website/src/components/Dashboard/Dashboard.tsx
@@ -17,13 +17,15 @@ const monthlyAttendance = [
 ];
 
 const AttendanceDashboard: React.FC = () => {
-  return ( 
-    <Box sx={{ flexGrow: 1, padding: 3, marginTop: {
-      xs: 2,     
-      sm: 3,     
-      md: 4,      
-      lg: 5,      
-    }, }}>
+  return (
+    <Box sx={{
+      flexGrow: 1, padding: 3, marginTop: {
+        xs: -5,
+        sm: -5,
+        md: -3,
+        lg: -3,
+      },
+    }}>
       <Grid container spacing={3}>
         {/* Attendance Summary */}
         <Grid item xs={12} md={4}>
